requirements.txt
django==2.2.4
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
Pillow==6.1.0
psycopg2==2.8.3

# dev dependancy 
pylint==2.3.1
pep8==1.7.1
autopep8==1.4.4
